feat: updated sim service with shared
