+++ libs/shared/services/src/lib/shared-services.ts
@@ -1,1 +1,3 @@
 export * from './format/format.service';
+export * from './shared-simulation/shared-simulation.service';
+export * from './shared-simulation-status/shared-simulation-status.service';
