+++ libs/shared/services/src/lib/shared-simulation-status/shared-simulation-status.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SharedSimulationStatusService } from './shared-simulation-status.service';
+
+describe('SharedSimulationStatusService', () => {
+  let service: SharedSimulationStatusService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SharedSimulationStatusService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
